progress create lending product

+++ lending/src/main/java/com/travel/lending/enitity/client/product/LendingProduct.java
@@ -2,6 +2,7 @@ package com.travel.lending.enitity.client.product;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,9 +12,9 @@ public class LendingProduct {
     private Long userIdBorrower;
     private String accountLender;
     private String accountBorrower;
-    private Long interest;
-    private Long amount;
-    private Long totalReimbursement;
+    private BigDecimal interest;
+    private BigDecimal amount;
+    private BigDecimal totalReimbursement;
     private LocalDateTime dueTime;
     private String reference;
     private String status;
